Refactoring App
- refactoring and restructuring app to have component scripts and separate css files
