+++ src/styles/App.css
@@ -1,0 +1,38 @@
+* {
+  font-size: 36px;
+}
+
+body {
+  margin: 0;
+  padding: 0;
+  font-family: sans-serif;
+  color: rgb(0, 0, 0);
+}
+
+.app-header {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  font-size: calc(10px + 1.5vmin);
+  border-bottom: 1px solid #000;
+  padding-bottom: 2px;
+}
+
+.app-background {
+  background-color: #eeecec;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  min-height: 100vh;
+}
+
+.main-container {
+  background-color: #ffffff;
+  max-width: 60%;
+  min-width: 500px;
+  height: min-content;
+  border-radius: 5px;
+  padding: 20px;
+  box-shadow: 10px 10px 26px 0px rgba(0, 0, 0, 0.35);
+}
