+++ src/styles/Todo.css
@@ -1,0 +1,44 @@
+* {
+  font-size: 36px;
+}
+
+.item-container {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  padding: 20px 0 20px 0;
+  border-bottom: 1px #000000 solid;
+}
+
+.item-name {
+  cursor: pointer;
+  display: flex;
+  flex-direction: column;
+  width: 80%;
+}
+
+.item-name svg {
+  margin-left: 5px;
+  margin-right: 5px;
+}
+
+.total {
+  float: right;
+  padding: 10px;
+}
+
+.quantity-box {
+  display: flex;
+  align-items: center;
+  border: 1px solid #ffffff;
+  background: #eeecec;
+  border-radius: 50px;
+  font-size: 12px;
+  color: #000000;
+  width: 175px;
+  justify-content: center;
+}
+
+.completed {
+  text-decoration: line-through;
+}
